fixed and reordered tests
