+++ tests/udp.rs
@@ -7,7 +7,10 @@ use std::{
     time::Duration,
 };
 
-use tokio::{sync::oneshot, time::timeout};
+use tokio::{
+    sync::oneshot,
+    time::{self, timeout},
+};
 use turmoil::{
     lookup,
     net::{self, UdpSocket},
@@ -26,10 +29,6 @@ async fn bind() -> std::result::Result<net::UdpSocket, std::io::Error> {
 
 async fn bind_to_v4(port: u16) -> std::result::Result<net::UdpSocket, std::io::Error> {
     net::UdpSocket::bind((IpAddr::from(Ipv4Addr::UNSPECIFIED), port)).await
-}
-
-async fn bind_to_v6(port: u16) -> std::result::Result<net::UdpSocket, std::io::Error> {
-    net::UdpSocket::bind((IpAddr::from(Ipv6Addr::UNSPECIFIED), port)).await
 }
 
 async fn send_ping(sock: &net::UdpSocket) -> Result<()> {
@@ -354,79 +353,182 @@ fn bulk_transfer() -> Result {
     sim.run()
 }
 
-// # IpVersion specific tests
-
 #[test]
-fn bind_ipv4_socket() -> Result {
-    let mut sim = Builder::new()
-        .ip_subnets(IpSubnets::from_iter([IpSubnet::V4(Ipv4Subnet::default())]))
-        .build();
-    sim.client("client", async move {
-        let sock = bind_to_v4(0).await?;
-        assert!(sock.local_addr().unwrap().is_ipv4());
+fn direct_ipv4_bind() -> Result {
+    let mut sim = Builder::new().build();
+    sim.node("alice").build_client(async {
+        let fail = UdpSocket::bind(":::0").await?;
+        let err = fail
+            .send_to(b"Never", "192.168.0.42:2000")
+            .await
+            .unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::InvalidInput);
+
+        time::sleep(Duration::from_secs(1)).await;
+        let udp = UdpSocket::bind("0.0.0.0:0").await?;
+        udp.send_to(b"Hello world", "192.168.0.42:2000").await?;
         Ok(())
     });
+
+    sim.node("bob")
+        .with_addr(Ipv4Addr::new(192, 168, 0, 42))
+        .build_client(async {
+            let udp = UdpSocket::bind("192.168.0.42:2000").await?;
+            let mut buf = [0; 128];
+            let (n, from) = udp.recv_from(&mut buf).await?;
+            assert_eq!(from.ip(), Ipv4Addr::new(192, 168, 0, 1));
+            assert_eq!(&buf[..n], b"Hello world");
+            Ok(())
+        });
+
+    sim.run()
+}
+
+fn double_ipv4_subnets() -> IpSubnets {
+    let mut subnets = IpSubnets::new();
+    subnets.add(IpSubnet::V4(Ipv4Subnet::new(
+        Ipv4Addr::new(192, 168, 0, 0),
+        16,
+    )));
+    subnets.add(IpSubnet::V4(Ipv4Subnet::new(
+        Ipv4Addr::new(10, 1, 0, 0),
+        16,
+    )));
+    subnets
+}
+
+#[test]
+fn direct_ipv6_bind() -> Result {
+    let mut sim = Builder::new().build();
+    sim.node("alice").build_client(async {
+        let fail = UdpSocket::bind("0.0.0.0:0").await?;
+        let err = fail.send_to(b"Never", "fe80::42:2000").await.unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::InvalidInput);
+
+        time::sleep(Duration::from_secs(1)).await;
+        let udp = UdpSocket::bind(":::0").await?;
+        udp.send_to(b"Hello world", "fe80::42:2000").await?;
+        Ok(())
+    });
+
+    sim.node("bob")
+        .with_addr(Ipv6Addr::new(0xfe80, 0, 0, 0, 0, 0, 0, 0x42))
+        .build_client(async {
+            let udp = UdpSocket::bind("fe80::42:2000").await?;
+            let mut buf = [0; 128];
+            let (n, from) = udp.recv_from(&mut buf).await?;
+            assert_eq!(from.ip(), Ipv6Addr::new(0xfe80, 0, 0, 0, 0, 0, 0, 1));
+            assert_eq!(&buf[..n], b"Hello world");
+            Ok(())
+        });
+
     sim.run()
 }
 
 #[test]
-fn bind_ipv6_socket() -> Result {
-    let mut sim = Builder::new()
-        .ip_subnets(IpSubnets::from_iter([IpSubnet::V6(Ipv6Subnet::default())]))
-        .build();
-    sim.client("client", async move {
-        let sock = bind_to_v6(0).await?;
-        assert!(sock.local_addr().unwrap().is_ipv6());
+fn interface_exclusivity() -> Result {
+    let mut sim = Builder::new().ip_subnets(double_ipv4_subnets()).build();
+    sim.node("bob").build_client(async {
+        let sock = UdpSocket::bind("192.168.0.1:2000").await?;
+        let mut buf = [0; 128];
+        let (n, _) = sock.recv_from(&mut buf).await?;
+        assert_eq!(&buf[..n], b"Hello");
         Ok(())
     });
+
+    sim.node("alice").build_client(async {
+        let fail = UdpSocket::bind("10.1.0.2:0").await?;
+        let err = fail
+            .send_to(b"Never", "192.168.0.1:2000")
+            .await
+            .unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::ConnectionRefused);
+
+        let succ = UdpSocket::bind("192.168.0.2:0").await?;
+        succ.send_to(b"Hello", "192.168.0.1:2000").await?;
+
+        Ok(())
+    });
+
     sim.run()
+}
+
+#[test]
+fn unspecified_bind_all_interface_receiver() -> Result {
+    let mut sim = Builder::new().ip_subnets(double_ipv4_subnets()).build();
+    sim.node("rx").build_client(async move {
+        let sock = UdpSocket::bind("0.0.0.0:2000").await?;
+        let mut buf = [0; 128];
+
+        let (n, origin) = sock.recv_from(&mut buf).await?;
+        assert_eq!(&buf[..n], b"Hello from TX 1");
+        assert_eq!(origin.ip(), Ipv4Addr::new(192, 168, 0, 2));
+
+        let (n, origin) = sock.recv_from(&mut buf).await?;
+        assert_eq!(&buf[..n], b"Hello from TX 2");
+        assert_eq!(origin.ip(), Ipv4Addr::new(10, 1, 0, 3));
+
+        Ok(())
+    });
+
+    sim.node("tx1").build_client(async move {
+        UdpSocket::bind("192.168.0.2:0")
+            .await?
+            .send_to(b"Hello from TX 1", "192.168.0.1:2000")
+            .await?;
+        Ok(())
+    });
+
+    sim.node("tx2").build_client(async move {
+        time::sleep(Duration::from_secs(1)).await;
+        UdpSocket::bind("10.1.0.3:0")
+            .await?
+            .send_to(b"Hello from TX 2", "10.1.0.1:2000")
+            .await?;
+        Ok(())
+    });
+
+    sim.run()
+}
+
+#[test]
+fn unspecified_bind_all_interfaces_sender() -> Result {
+    let mut sim = Builder::new().ip_subnets(double_ipv4_subnets()).build();
+    sim.node("rx1")
+        .build_client(rx_once("192.168.0.1:2000", "Hello to RX 1"));
+    sim.node("rx2")
+        .build_client(rx_once("10.1.0.2:2000", "Hello to RX 2"));
+
+    sim.node("tx").build_client(async move {
+        let sock = UdpSocket::bind("0.0.0.0:0").await?;
+        sock.send_to(b"Hello to RX 1", "192.168.0.1:2000").await?;
+        sock.send_to(b"Hello to RX 2", "10.1.0.2:2000").await?;
+        Ok(())
+    });
+
+    sim.run()
+}
+
+async fn rx_once(bind: &str, expect: &str) -> Result<()> {
+    let sock = UdpSocket::bind(bind).await?;
+    let mut buf = [0; 128];
+    let (n, _) = sock.recv_from(&mut buf).await?;
+    assert_eq!(&buf[..n], expect.as_bytes());
+    Ok(())
 }
 
 #[test]
 #[should_panic]
-fn bind_ipv4_version_missmatch() {
-    let mut sim = Builder::new()
-        .ip_subnets(IpSubnets::from_iter([IpSubnet::V6(Ipv6Subnet::default())]))
-        .build();
-    sim.client("client", async move {
-        let _sock = bind_to_v4(0).await?;
-        Ok(())
-    });
-    sim.run().unwrap()
-}
+fn direct_bind_ip_version_not_supported() {
+    let mut subnets = IpSubnets::new();
+    subnets.add(IpSubnet::V4(Ipv4Subnet::new(Ipv4Addr::new(1, 2, 3, 0), 24)));
+    let mut sim = Builder::new().ip_subnets(subnets).build();
+    sim.node("alice").build_client(async {
+        let _ = UdpSocket::bind(":::0").await?;
 
-#[test]
-#[should_panic]
-fn bind_ipv6_version_missmatch() {
-    let mut sim = Builder::new()
-        .ip_subnets(IpSubnets::from_iter([IpSubnet::V4(Ipv4Subnet::default())]))
-        .build();
-    sim.client("client", async move {
-        let _sock = bind_to_v6(0).await?;
         Ok(())
     });
-    sim.run().unwrap()
-}
-
-#[test]
-fn ipv6_connectivity() -> Result {
-    let mut sim = Builder::new()
-        .ip_subnets(IpSubnets::from_iter([IpSubnet::V6(Ipv6Subnet::default())]))
-        .build();
-    sim.client("server", async move {
-        let sock = UdpSocket::bind(":::80").await.unwrap();
-        let mut buf = [0; 512];
-        let _stream = sock.recv_from(&mut buf).await.unwrap();
-        Ok(())
-    });
-    sim.client("client", async move {
-        let sock = UdpSocket::bind(":::0").await.unwrap();
-        sock.send_to(&[1], "server:80").await.unwrap();
-        let _ = sock;
-        Ok(())
-    });
-
-    sim.run()
+    sim.run().expect("paniced at subtask");
 }
 
 #[test]
